requirements.txt
Django>=4.2.13,<4.3
djangorestframework>=3.15.1,<3.16
psycopg2>=2.9.9,<2.10
drf-spectacular>=0.27.2,<0.28
Pillow>=10.3.0,<10.4.0
django-cors-headers>=4.3.1,<4.4
djangorestframework-simplejwt>=5.3.1,<5.4
PyJWT>=2.8.0,<2.9
